+++ src/main/java/com/dhanantry/scapeandrunparasites/client/renderer/entity/ancient/RenderAncientPod.java
@@ -1,12 +1,3 @@
-/*
- * Decompiled with CFR 0.152.
- * 
- * Could not load the following classes:
- *  net.minecraft.client.renderer.GlStateManager
- *  net.minecraft.client.renderer.entity.RenderManager
- *  net.minecraft.util.ResourceLocation
- *  net.minecraft.util.math.MathHelper
- */
 package com.dhanantry.scapeandrunparasites.client.renderer.entity.ancient;
 
 import com.dhanantry.scapeandrunparasites.client.model.entity.projectile.ModelDropPod;
@@ -17,27 +8,25 @@ import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.MathHelper;
 
-public class RenderAncientPod
-extends RenderSRP<EntityDropPod> {
-    public static final ResourceLocation TEXTURES = new ResourceLocation("srparasites:textures/entity/monster/ancientpod.png");
+public class RenderAncientPod extends RenderSRP<EntityDropPod> {
+   public static final ResourceLocation TEXTURES = new ResourceLocation("srparasites:textures/entity/monster/ancientpod.png");
 
-    public RenderAncientPod(RenderManager manager) {
-        super(manager, new ModelDropPod(), 0.8f);
-    }
+   public RenderAncientPod(RenderManager manager) {
+      super(manager, new ModelDropPod(), 0.8F);
+   }
 
-    protected void preRenderCallback(EntityDropPod entitylivingbaseIn, float partialTickTime) {
-        float f = entitylivingbaseIn.getSelfeFlashIntensity(partialTickTime);
-        float f1 = 1.0f + MathHelper.func_76126_a((float)(f * 100.0f)) * f * 0.01f;
-        f = MathHelper.func_76131_a((float)f, (float)0.0f, (float)1.0f);
-        f *= f;
-        f *= f;
-        float f2 = (1.0f + f * 0.4f) * f1;
-        float f3 = (1.0f + f * 0.1f) / f1;
-        GlStateManager.func_179152_a((float)f2, (float)f3, (float)f2);
-    }
+   protected void preRenderCallback(EntityDropPod entitylivingbaseIn, float partialTickTime) {
+      float f = entitylivingbaseIn.getSelfeFlashIntensity(partialTickTime);
+      float f1 = 1.0F + MathHelper.func_76126_a(f * 100.0F) * f * 0.01F;
+      f = MathHelper.func_76131_a(f, 0.0F, 1.0F);
+      f *= f;
+      f *= f;
+      float f2 = (1.0F + f * 0.4F) * f1;
+      float f3 = (1.0F + f * 0.1F) / f1;
+      GlStateManager.func_179152_a(f2, f3, f2);
+   }
 
-    protected ResourceLocation getEntityTexture(EntityDropPod entity) {
-        return TEXTURES;
-    }
+   protected ResourceLocation getEntityTexture(EntityDropPod entity) {
+      return TEXTURES;
+   }
 }
-
